+++ src/table/CustomScrollbars.tsx
@@ -2,15 +2,26 @@ import * as React from 'react';
 import Scrollbars from 'react-custom-scrollbars';
 
 export interface CustomScrollbarsProps {
-    style: React.CSSProperties;
+    style?: React.CSSProperties;
     onScroll?: any;
-    children?: React.ReactChildren;
+    forwardedRef: (elem: Element) => void;
+    children?: React.ReactNode;
 }
 
-export class CustomScrollbars extends React.PureComponent<CustomScrollbarsProps> {
+class CustomScrollbarsInner extends React.PureComponent<CustomScrollbarsProps> {
     public render() {
+
+        // https://github.com/bvaughn/react-window/issues/110#issuecomment-469061213
+
+        const refSetter = (elem: any) => {
+            if (!this.props.forwardedRef)
+                return;
+            this.props.forwardedRef((elem || {}).view);
+        };
+
         return (
             <Scrollbars
+                ref={refSetter}
                 style={{
                     ...this.props.style,
                     overflow: 'hidden',
@@ -23,3 +34,8 @@ export class CustomScrollbars extends React.PureComponent<CustomScrollbarsProps>
         );
     }
 }
+
+export const CustomScrollbars = React.forwardRef((props, ref) => (
+    <CustomScrollbarsInner {...props} forwardedRef={ref as any} />
+));
+CustomScrollbars.displayName = 'CustomScrollbars';