+++ src/utils/index.ts
@@ -1,4 +1,5 @@
 export * from './ErrorBoundary';
+export * from './keys';
 export * from './reactUtils';
 export * from './sizeUtils';
 export * from './utils';