+++ src/utils/keys.ts
@@ -1,0 +1,6 @@
+export enum Keys {
+    PageUp = 'PageUp',
+    PageDown = 'PageDown',
+    End ='End',
+    Home ='Home',
+}